pyproject.toml
[project]
name = "pixelpanda-mcp"
version = "0.1.0"
description = "MCP server for PixelPanda — AI image processing tools for Claude Desktop, Cursor, and other MCP clients"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pillow>=10.0.0",
    "httpx>=0.25.0",
]

[project.scripts]
pixelpanda-mcp = "pixelpanda_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pixelpanda_mcp"]
